PZW_Projket/requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.0.6
WTForms==3.2.1

